igraph::linalg

Function lapack_dgetrs

pub fn lapack_dgetrs(
    transpose: bool,
    lu: &Matrix,
    ipiv: &[i64],
    b: &Matrix,
) -> Result<Matrix>

Solves A X = B or A' X = B (transpose) using the LU factors of a square A computed by lapack_dgetrf (igraph_lapack_dgetrs). No check is made for singularity.

Binds igraph_lapack_dgetrs.

Errors

If lu is not square, b has the wrong number of rows, or the pivots are out of range.